fix: 복잡한 문자열 표현에 대하여 String.format 사용

+++ src/main/java/org/prgms/voucheradmin/domain/voucher/dao/JdbcVoucherRepository.java
@@ -109,7 +109,7 @@ public class JdbcVoucherRepository implements VoucherRepository{
                 long amount = resultSet.getLong("voucher_amount");
                 return new FixedAmountVoucher(voucherId, amount);
             default:
-                long percent = resultSet.getLong("voucher_amount");
+                int percent = resultSet.getInt("voucher_amount");
                 return new PercentageDiscountVoucher(voucherId, percent);
         }
     };
